Restyled by prettier

+++ index.js
@@ -1,10 +1,10 @@
 require('dotenv').config();
-const {Client, GatewayIntentBits} = require('discord.js');
+const { Client, GatewayIntentBits } = require('discord.js');
 
 const DiscordClient = new Client({
-  intents : GatewayIntentBits.Guilds,
-  messageCacheMaxSize : 0,
-  disableMentions : 'everyone'
+  intents: GatewayIntentBits.Guilds,
+  messageCacheMaxSize: 0,
+  disableMentions: 'everyone',
 });
 
 const DelayBetween = 60 * 1000;
@@ -19,8 +19,7 @@ const TwoWeekOffset = 14 * 24 * 60 * 60 * 1000;
 const ChannelConfigs = new Map();
 
 function ChannelName(channel) {
-  return channel.guild != null ? channel.guild.name + ' #' + channel.name
-                               : '#' + channel.name;
+  return channel.guild != null ? channel.guild.name + ' #' + channel.name : '#' + channel.name;
 }
 function Log(msg) {
   console.log(new Date().toTimeString().substring(0, 9) + msg);
@@ -32,47 +31,34 @@ async function Wipe(channelConfig, reWipe) {
 
   try {
     let now = Date.now();
-    let wipeSnowstamp =
-        (BigInt(now - 14200704e5 /*DISCORD_EPOCH*/ - channelConfig.ttl) << 22n)
-            .toString();
+    let wipeSnowstamp = (BigInt(now - 14200704e5 /*DISCORD_EPOCH*/ - channelConfig.ttl) << 22n).toString();
     let twoWeeksAgo = now - TwoWeekOffset;
-    let messages = await channel.messages.fetch(
-        {limit : 100, before : wipeSnowstamp}, false);
-    messages = messages.filter(
-        (message) => message.createdTimestamp >
-                     twoWeeksAgo); // because before and after are mutually
-                                   // exclusive in the query
+    let messages = await channel.messages.fetch({ limit: 100, before: wipeSnowstamp }, false);
+    messages = messages.filter((message) => message.createdTimestamp > twoWeeksAgo); // because before and after are mutually
+    // exclusive in the query
     let fetchSize = messages.size;
     messages = messages.filter((message) => !message.pinned);
     if (messages.size !== 0) {
       Log(`Deleting ${messages.size} messages in ${ChannelName(channel)}`);
       await channel.bulkDelete(messages);
 
-      if (interval > DayMs)
-        channelConfig.int = interval = DayMs; // in case it was extended before
+      if (interval > DayMs) channelConfig.int = interval = DayMs; // in case it was extended before
 
       if (fetchSize === 100) {
         channelConfig.t = setTimeout(Wipe, DelayBetween, channelConfig, true);
-        if (interval > MinInterval)
-          channelConfig.int = Math.min(interval / 2, MinInterval);
+        if (interval > MinInterval) channelConfig.int = Math.min(interval / 2, MinInterval);
         return;
       }
-      if (interval < DayMs)
-        channelConfig.int = interval = Math.max(interval * 2, DayMs);
+      if (interval < DayMs) channelConfig.int = interval = Math.max(interval * 2, DayMs);
     } else if (!reWipe) {
       if (channelConfig.ttl < /*6 days*/ 518400000) {
         if (interval < /*6 days*/ 518400000) {
           channelConfig.int = interval += IntervalExtension;
         }
       } else if (interval >= DayMs) {
-        let messages = await channel.messages.fetch(
-            {limit : 1, after : wipeSnowstamp}, false);
-        if (messages.size !== 0)
-          interval = Math.max(messages.first().createdTimestamp +
-                                  channelConfig.tll - now,
-                              DayMs);
-        else
-          interval = channelConfig.ttl + DayMs;
+        let messages = await channel.messages.fetch({ limit: 1, after: wipeSnowstamp }, false);
+        if (messages.size !== 0) interval = Math.max(messages.first().createdTimestamp + channelConfig.tll - now, DayMs);
+        else interval = channelConfig.ttl + DayMs;
       }
     }
   } catch (err) {
@@ -83,12 +69,10 @@ async function Wipe(channelConfig, reWipe) {
       // Post to channel that the bot is lacking required permissions.
       channel.send(`${ChannelName(channel)}: ${err.message}`);
       return;
-    } else
-      console.error(err);
+    } else console.error(err);
   }
 
-  channelConfig.t =
-      setTimeout(Wipe, interval + Math.random * RecurringSpread, channelConfig);
+  channelConfig.t = setTimeout(Wipe, interval + Math.random * RecurringSpread, channelConfig);
 }
 
 let MyId;
@@ -96,32 +80,25 @@ let MyId;
 async function DeleteOldAnnounce(channel) {
   try {
     let pins = await channel.messages.fetchPinned(false);
-    let announcePin = pins.find((message) => message.author != null &&
-                                             message.author.id === MyId);
-    if (announcePin != null)
-      await announcePin.delete();
-  } catch (err) {
-  }
+    let announcePin = pins.find((message) => message.author != null && message.author.id === MyId);
+    if (announcePin != null) await announcePin.delete();
+  } catch (err) {}
 }
 
 function AddChannel(matches, channel, announce) {
-  let param = matches.length === 1
-                  ? matches[0].substring(9).trimStart()
-                  : matches.map((x) => x.substring(9).trimStart())
-                        .sort((a, b) => b.length - a.length)[0];
+  let param = matches.length === 1 ? matches[0].substring(9).trimStart() : matches.map((x) => x.substring(9).trimStart()).sort((a, b) => b.length - a.length)[0];
   param = param === '' ? 7 : Math.min(12, Math.max(param, 1));
-  let channelConfig = {ttl : param * DayMs, int : DayMs, channel};
+  let channelConfig = { ttl: param * DayMs, int: DayMs, channel };
   ChannelConfigs.set(channel.id, channelConfig);
-  channelConfig.t =
-      setTimeout(Wipe, Math.random * InitialSpread, channelConfig, true);
+  channelConfig.t = setTimeout(Wipe, Math.random * InitialSpread, channelConfig, true);
   Log(`Adding ${ChannelName(channel)} with a ${param} day wipe`);
   if (announce) {
-    DeleteOldAnnounce(channel).then(
-        () => channel
-                  .send('The messages will be deleted after ' +
-                        (param === 1 ? 'one day' : param + ' days'))
-                  .then((message) => message.pin().catch())
-                  .catch());
+    DeleteOldAnnounce(channel).then(() =>
+      channel
+        .send('The messages will be deleted after ' + (param === 1 ? 'one day' : param + ' days'))
+        .then((message) => message.pin().catch())
+        .catch()
+    );
   }
 }
 
@@ -152,7 +129,7 @@ function ProcessGuild(guild) {
 
 DiscordClient.on('ready', () => {
   Log(`Logged in as ${DiscordClient.user.tag}!`);
-  DiscordClient.user.setActivity('Deleting messages', {type : 'WATCHING'});
+  DiscordClient.user.setActivity('Deleting messages', { type: 'WATCHING' });
   MyId = DiscordClient.user.id;
   DiscordClient.on('error', console.error);
 
@@ -162,23 +139,19 @@ DiscordClient.on('ready', () => {
 });
 
 DiscordClient.on('channelDelete', (channel) => {
-  if (channel.type !== 'voice' && channel.type !== 'category')
-    DeleteChannel(channel.id);
+  if (channel.type !== 'voice' && channel.type !== 'category') DeleteChannel(channel.id);
 });
 
 DiscordClient.on('guildDelete', (guild) => {
   for (let channel of guild.channels.cache.values()) {
-    if (channel.type !== 'voice' && channel.type !== 'category')
-      DeleteChannel(channel.id);
+    if (channel.type !== 'voice' && channel.type !== 'category') DeleteChannel(channel.id);
   }
 });
 
 DiscordClient.on('guildCreate', ProcessGuild);
 
 DiscordClient.on('channelUpdate', async (oldChannel, newChannel) => {
-  if (newChannel.type === 'voice' || newChannel.type === 'category' ||
-      oldChannel.topic === newChannel.topic)
-    return;
+  if (newChannel.type === 'voice' || newChannel.type === 'category' || oldChannel.topic === newChannel.topic) return;
 
   let announceDelete = DeleteChannel(newChannel.id);
 
@@ -192,8 +165,7 @@ DiscordClient.on('channelUpdate', async (oldChannel, newChannel) => {
   }
 
   if (announceDelete) {
-    newChannel.send('The messages will not be automatically deleted')
-        .catch(console.error);
+    newChannel.send('The messages will not be automatically deleted').catch(console.error);
     await DeleteOldAnnounce(newChannel);
   }
 });
@@ -201,12 +173,13 @@ DiscordClient.on('rateLimit', (rateLimitInfo) => {
   if (rateLimitInfo.global) {
     let time = rateLimitInfo.reset * 1000;
     Log(`Rate limit exceeded. Reset at ${new Date(time).toLocaleString()}`);
-    setTimeout(() => { Log('Rate limit cleared, resuming.'); },
-               time - Date.now());
+    setTimeout(() => {
+      Log('Rate limit cleared, resuming.');
+    }, time - Date.now());
   }
 });
-const dbl = (require('dblapi.js'));
+const dbl = require('dblapi.js');
 new dbl(process.env.DBLTOKEN, DiscordClient);
 DiscordClient.login(process.env.TOKEN)
-    .then(() => Log(`Logged in as ${DiscordClient.user.tag}!`))
-    .catch(console.error);
+  .then(() => Log(`Logged in as ${DiscordClient.user.tag}!`))
+  .catch(console.error);
